+++ nativelink-store/tests/redis_store_test.rs
@@ -237,7 +237,7 @@ async fn upload_and_get_data() -> Result<(), Error> {
     );
 
     let result = store
-        .get_part_unchunked(digest, 0, Some(data.len()))
+        .get_part_unchunked(digest, 0, Some(data.len() as u64))
         .await
         .unwrap();
 
@@ -318,7 +318,7 @@ async fn upload_and_get_data_with_prefix() -> Result<(), Error> {
     );
 
     let result = store
-        .get_part_unchunked(digest, 0, Some(data.len()))
+        .get_part_unchunked(digest, 0, Some(data.len() as u64))
         .await
         .unwrap();
 
@@ -462,7 +462,7 @@ async fn test_large_downloads_are_chunked() -> Result<(), Error> {
     );
 
     let get_result: Bytes = store
-        .get_part_unchunked(digest, 0, Some(data.clone().len()))
+        .get_part_unchunked(digest, 0, Some(data.clone().len() as u64))
         .await
         .unwrap();
 
@@ -553,7 +553,7 @@ async fn yield_between_sending_packets_in_update() -> Result<(), Error> {
     tokio::try_join!(
         async {
             store
-                .update(digest, rx, UploadSizeInfo::ExactSize(data.len()))
+                .update(digest, rx, UploadSizeInfo::ExactSize(data.len() as u64))
                 .await
                 .unwrap();
 
@@ -576,7 +576,7 @@ async fn yield_between_sending_packets_in_update() -> Result<(), Error> {
     );
 
     let result = store
-        .get_part_unchunked(digest, 0, Some(data.clone().len()))
+        .get_part_unchunked(digest, 0, Some(data.clone().len() as u64))
         .await
         .unwrap();
